Start work, basic setup
Established basic GUI look using a previous project (my own project).
This is by no means usable yet.

Created Grid and Tile classes, in order to simplify further work and are yet to be entirely completed.
These classes should be used to describe the map area, that will be edited with the use of buttons from GUI.

The GUI will work with buttons, which will display the Tile data, which will then be exported to a .json file.