GUI changes (heigth&width TF and Grid Resize Button)
Added TextFields for manual setting for Grid size and a Button to create the grid.
Currently, the TextFields are just there visually.
The Button spawns new Grid Tiles, but does not remove the old ones. This will be implemented next.
Changing the Grid size will remove all previous changes to the map, so be careful.