Implement ScrollPane
The Grid is now displayed as Button Tiles on a Scroll Pane, so the Grid can be of any size and still displayed.

The problem that remains is the size if the Tiles. They take as much space as possible, but this leads to unsusable non-square rectangles as the difference between the number of Grid columns and rows increases.