Tile size retry
Tiles depend on the size of the JFrame. They should be 50x50 if the number of tiles is large enough. For the default size of the JFrame, if the dimensions are at least 12 or so, the scroll bars appear and the Tiles are set to 50x50size.
Most likely the number 12 is due to the fact that JFrame - settings pannel is about 600x600 and Tile size is 50x50.
Just don't make height/width <10 and the program can be used normaly.