Implement Edit Window (orientation only)
When the Edit button is clicked, a new window now opens, in which the user can set the direction of the selected Tiles, by setting the ComboBox and clicking the Apply button.
The change is displayed in real time on the previous window, with the orientation of the tiles being represented by a green arrow, pointing to the seleced direction.