Implement isRoad editing
isRoad value of a Tile can now be changed through the editor similarly to orientation. The change is displayed through the change of the background of the tile (Black for building, white for road). And since the Icon's background is translucent, both the orientation and state of the Tile can be viewed at the same time.