Write basic Grid information to JSON
JsonMaker now writes all currently implemented information about the Grid Tiles into a JSON file.

There might be an issue however. at the end of each JSONObject within the JSON file, there is a trailing comma at the end, which might pose a problem in the future.
Will try to remove it, if it is a real issue, as well as having the JSON display without whitespace.