Add filename TextField functionality and failsafes
When creating the Json file for the map, the user can now add the name of the file (without .json extension) and a .json file with that name will be created (or overwritten).
If such a file cannot be created, or if there are any other exceptions, the program will write to a default file Map.json.