Comment changes to distinguish parts of code
Code in EditGUI was getting crowded, since all of the components are made in a single method.
Because of that, comments for different sections of the EditGUI have been written.
This is a temporary solution and should most likely be replaced by classes that extend JPanel in order to make it optimal for reading.