Barricade class Type only
Barricade class now only uses the type attribute, which has been changed to String as opposed to int,
since there should not be too many options at the start and can be implemented in the MapReader instead.

The array of possible types should be an array of Strings, hardcoded (later editable) in the Grid class.