FRONTEND: Start on Mesh - Demo cases
Started frontend part of the applying Mesh/MeshIndex to Tiles. A TODO has been added to make the lists for the combobox dynamic and based on the current state of the source folder.
For now, default list has been made until a dynamic version as well as actuall meshes are made as well.