BACKEND: Apply mesh_idx button implementation
The current version for the ActionListener for the apply mesh_idx button has been implemented. This has to be changed later tho, so that Building meshes cant be applied to Roads and vice versa. The indexes might also be set to pure file name Strings in the future as well.